Extract out emailCapabilities into tests
